api/requirements.txt
connexion == 2.0.2
swagger-ui-bundle == 0.0.2
python_dateutil == 2.6.0
setuptools >= 21.0.0
flask-cors >= 3.0.7
